# Implement block_given? call as optimized instruction