--- a/core/src/main/java/org/jruby/ir/builder/IRBuilderAST.java
+++ b/core/src/main/java/org/jruby/ir/builder/IRBuilderAST.java
@@ -1301,7 +1301,7 @@ public class IRBuilderAST extends IRBuilder<Node, DefNode, WhenNode, RescueBodyN
                     )
             );
         case YIELDNODE:
-            return buildDefinitionCheck(new BlockGivenInstr(temp(), getYieldClosureVariable()), DefinedMessage.YIELD.getText());
+            return buildDefinitionCheck(new BlockGivenInstr(temp(), getYieldClosureVariable(), true), DefinedMessage.YIELD.getText());
         case ZSUPERNODE:
             return addResultInstr(
                     new RuntimeHelperCall(
@@ -2034,6 +2034,14 @@ public class IRBuilderAST extends IRBuilder<Node, DefNode, WhenNode, RescueBodyN
     public Operand buildFCall(Variable result, FCallNode node) {
         if (result == null) result = temp();
         RubySymbol name = methodName = node.getName();
+
+        // special case methods with frame handling
+        if (name.idString().equals("block_given?")
+                && node.getArgsNode() == null
+                && node.getIterNode() == null) {
+            addInstr(new BlockGivenInstr(result, getYieldClosureVariable(), false));
+            return result;
+        }
 
         return createCall(result, buildSelf(), FUNCTIONAL, name, node.getArgsNode(), node.getIterNode(),
                 node.getLine(), node.isNewline());