--- a/core/src/main/java/org/jruby/ir/instructions/BlockGivenInstr.java
+++ b/core/src/main/java/org/jruby/ir/instructions/BlockGivenInstr.java
@@ -12,34 +12,60 @@ import org.jruby.parser.StaticScope;
 import org.jruby.runtime.DynamicScope;
 import org.jruby.runtime.ThreadContext;
 import org.jruby.runtime.builtin.IRubyObject;
+import org.jruby.runtime.callsite.FunctionalCachingCallSite;
 
 /**
  * Represents a defined?(yield) check, which works like a call to block_given? without
  * requiring special access to the caller's frame.
  */
 public class BlockGivenInstr extends OneOperandResultBaseInstr implements FixedArityInstr {
-    public BlockGivenInstr(Variable result, Operand block) {
+    private final boolean defined;
+    private final FunctionalCachingCallSite blockGivenSite;
+
+    public BlockGivenInstr(Variable result, Operand block, boolean defined) {
         super(Operation.BLOCK_GIVEN, result, block);
 
         assert result != null: "BlockGivenInstr result is null";
+
+        this.defined = defined;
+
+        if (defined) {
+            blockGivenSite = null;
+        } else {
+            blockGivenSite = new FunctionalCachingCallSite("block_given?");
+        }
     }
 
     public Operand getBlockArg() {
         return getOperand1();
     }
 
+    public boolean isDefined() {
+        return defined;
+    }
+
     @Override
     public Instr clone(CloneInfo ii) {
-        return new BlockGivenInstr(ii.getRenamedVariable(result), getBlockArg().cloneForInlining(ii));
+        return new BlockGivenInstr(ii.getRenamedVariable(result), getBlockArg().cloneForInlining(ii), defined);
     }
 
     public static BlockGivenInstr decode(IRReaderDecoder d) {
-        return new BlockGivenInstr(d.decodeVariable(), d.decodeOperand());
+        return new BlockGivenInstr(d.decodeVariable(), d.decodeOperand(), d.decodeBoolean());
+    }
+
+    @Override
+    public void encode(IRWriterEncoder e) {
+        super.encode(e);
+        e.encode(defined);
     }
 
     @Override
     public Object interpret(ThreadContext context, StaticScope currScope, DynamicScope currDynScope, IRubyObject self, Object[] temp) {
         Object blk = getBlockArg().retrieve(context, self, currScope, currDynScope, temp);
+
+        if (!defined) {
+            return IRRuntimeHelpers.blockGivenOrCall(context, self, blockGivenSite, blk);
+        }
 
         return IRRuntimeHelpers.isBlockGiven(context, blk);
     }