--- a/core/src/main/java/org/jruby/ir/targets/InvocationCompiler.java
+++ b/core/src/main/java/org/jruby/ir/targets/InvocationCompiler.java
@@ -131,8 +131,6 @@ public interface InvocationCompiler {
 
     /**
      * Invoke block_given? with awareness of any built-in methods.
-     *
-     * @param call a CallBase representing the call to block_given?
      */
-    void invokeBlockGiven(String file, String scopeFieldName, CallBase call);
+    void invokeBlockGiven(String file, String scopeFieldName);
 }