--- a/core/src/main/java/org/jruby/ir/targets/InvocationCompiler.java
+++ b/core/src/main/java/org/jruby/ir/targets/InvocationCompiler.java
@@ -132,5 +132,5 @@ public interface InvocationCompiler {
     /**
      * Invoke block_given? or iterator? with awareness of any built-in methods.
      */
-    void invokeBlockGiven(String callName, String file);
+    void invokeBlockGiven(String methodName, String file);
 }