--- a/core/src/main/java/org/jruby/ir/targets/JVMVisitor.java
+++ b/core/src/main/java/org/jruby/ir/targets/JVMVisitor.java
@@ -1281,6 +1281,18 @@ public class JVMVisitor extends IRVisitor {
                 throw new NotCompilableException("ruby2_keywords can change behavior of already-compiled code");
         }
 
+        // known frame-aware methods get compiled with special call sites that pass frame data through
+        switch (call.getName().idString()) {
+            case "block_given?":
+                if (call.getArgsCount() == 0 && !call.isPotentiallyRefined() && call.getClosureArg() == NullBlock.INSTANCE) {
+                    /* likely a call to built-in block_given? which just checks the frame block isGiven. Only optimized
+                       for method scopes currently. */
+                    m.getInvocationCompiler().invokeBlockGiven(file, jvm.methodData().scopeField, call);
+                    handleCallResult(m, call);
+                    return;
+                }
+        }
+
         boolean functional = call.getCallType() == CallType.FUNCTIONAL || call.getCallType() == CallType.VARIABLE;
 
         Operand[] args = call.getCallArgs();
@@ -1323,6 +1335,10 @@ public class JVMVisitor extends IRVisitor {
                 break;
         }
 
+        handleCallResult(m, call);
+    }
+
+    private void handleCallResult(IRBytecodeAdapter m, CallBase call) {
         Variable result = call.getResult();
         if (result != null) {
             if (!omitStoreLoad) jvmStoreLocal(result);