… bench/bench_block_given.rb
… core/src/main/java/org/jruby/ir/targets/indy/Bootstrap.java
… core/src/main/java/org/jruby/ir/targets/simple/NormalInvocationCompiler.java