This PR reworks fcalls to `block_given?` as a custom instruction based on `BlockGivenInstr` used by the lighter-weight `defined?(yield)` form.

Instead of forcing a caller frame (so that the caller's received block can be accessed downstack), the new logic checks if the target method is the core version and uses `defined?(yield)` logic in that case. When the method is not from JRuby core, then we dispatch normally... but without the caller frame requirement, since only the core `block_given?` is allowed to have that access. All other call paths leading to `block_given?` continue to use the deoptimized logic, so we are not introducing any new incompatibility.

This makes the performance of `block_given?`-calling methods equivalent to those that use `defined?(yield)` and avoids the framing cost in all situations that fcall `block_given?`.

Performance is shown through the added benchmark:

```
[] jruby $ (chruby jruby-9.4.5.0 ; jruby bench/bench_block_given.rb)
Warming up --------------------------------------
               block   417.419k i/100ms
            no block   609.148k i/100ms
       defined block   714.207k i/100ms
    defined no block     1.711M i/100ms
Calculating -------------------------------------
               block      4.199M (± 0.5%) i/s -     21.288M in   5.069484s
            no block      6.514M (± 0.3%) i/s -     32.894M in   5.049702s
       defined block      7.291M (± 0.5%) i/s -     37.139M in   5.093778s
    defined no block     17.039M (± 0.4%) i/s -     85.566M in   5.021866s
[] jruby $ jruby bench/bench_block_given.rb                         
Warming up --------------------------------------
               block   713.008k i/100ms
            no block     1.818M i/100ms
       defined block   752.260k i/100ms
    defined no block     1.709M i/100ms
Calculating -------------------------------------
               block      7.670M (± 0.4%) i/s -     38.502M in   5.020222s
            no block     18.201M (± 0.4%) i/s -     92.738M in   5.095348s
       defined block      7.569M (± 0.6%) i/s -     38.365M in   5.068610s
    defined no block     17.051M (± 1.3%) i/s -     85.467M in   5.013348s
```